requirements.txt
Django>=2.2,<3
djangorestframework>=3
psycopg2>=2.8.5